+++ src/pluginHooks.js
@@ -13,18 +13,39 @@ Handsontable.PluginHooks = {
     walkontableConfig: []
   },
 
-  push: function (hook, fn) {
-    this.hooks[hook].push(fn);
+  push: function (key, fn) {
+    this.hooks[key].push(fn);
   },
 
-  unshift: function (hook, fn) {
-    this.hooks[hook].unshift(fn);
+  unshift: function (key, fn) {
+    this.hooks[key].unshift(fn);
   },
 
-  run: function (instance, hook, p1, p2, p3, p4, p5) {
+  run: function (instance, key, p1, p2, p3, p4, p5) {
     //performance considerations - http://jsperf.com/call-vs-apply-for-a-plugin-architecture
-    for (var i = 0, ilen = this.hooks[hook].length; i < ilen; i++) {
-      this.hooks[hook][i].call(instance, p1, p2, p3, p4, p5);
+    for (var i = 0, ilen = this.hooks[key].length; i < ilen; i++) {
+      this.hooks[key][i].call(instance, p1, p2, p3, p4, p5);
     }
+  }
+};
+
+Handsontable.PluginModifiers = {
+  modifiers: {
+    col: []
+  },
+
+  push: function (key, fn) {
+    this.modifiers[key].push(fn);
+  },
+
+  unshift: function (key, fn) {
+    this.modifiers[key].unshift(fn);
+  },
+
+  run: function (instance, key, p1, p2, p3, p4, p5) {
+    for (var i = 0, ilen = this.modifiers[key].length; i < ilen; i++) {
+      p1 = this.modifiers[key][i].call(instance, p1, p2, p3, p4, p5);
+    }
+    return p1;
   }
 };